BAFK-266 Initial commit of refactors to codebase, breaking up utils and renaming some classes.
