+++ src/main/java/com/bounteous/aem/compgenerator/AemCompGenerator.java
@@ -21,8 +21,7 @@ import com.bounteous.aem.compgenerator.exceptions.GeneratorException;
 import com.bounteous.aem.compgenerator.javacodemodel.JavaCodeModel;
 import com.bounteous.aem.compgenerator.models.GenerationConfig;
 import com.bounteous.aem.compgenerator.utils.CommonUtils;
-import com.bounteous.aem.compgenerator.utils.ComponentGeneratorUtils;
-import org.apache.commons.lang.StringUtils;
+import com.bounteous.aem.compgenerator.utils.ComponentUtils;
 
 import java.io.File;
 
@@ -35,7 +34,10 @@ import java.io.File;
  */
 public class AemCompGenerator {
 
-    public static void main(String[] args) {
+    private GenerationConfig config;
+
+
+    public void main(String[] args) {
         try {
             String configPath = "data-config.json";
             if (args != null && args.length > 0) {
@@ -48,14 +50,14 @@ public class AemCompGenerator {
                 throw new GeneratorException("Config file missing / empty.");
             }
 
-            GenerationConfig config = CommonUtils.getComponentData(configFile);
+            config = CommonUtils.getComponentData(configFile);
 
+            //TODO: move these checks to CommonUtils
             if (config == null) {
                 throw new GeneratorException("Config file is empty / null !!");
             }
 
-            if (StringUtils.isBlank(config.getName()) || StringUtils.isBlank(config.getType())
-                    || config.getProjectSettings() == null || StringUtils.isBlank(config.getProjectSettings().getComponentPath())) {
+            if (!config.isValid() || !CommonUtils.isModelValid(config.getProjectSettings())) {
                 throw new GeneratorException("Mandatory fields missing in the data-config.json !");
             }
 
@@ -65,7 +67,7 @@ public class AemCompGenerator {
             config.setCompDir(compDir);
 
             //builds component folder and file structure.
-            ComponentGeneratorUtils generatorUtils = new ComponentGeneratorUtils(config);
+            ComponentUtils generatorUtils = new ComponentUtils(config);
             generatorUtils._buildComponent();
 
             //builds sling model based on config.
