+++ src/main/java/com/bounteous/aem/compgenerator/models/BaseModel.java
@@ -1,0 +1,6 @@
+package com.bounteous.aem.compgenerator.models;
+
+public interface BaseModel {
+
+    boolean isValid();
+}
