+++ src/main/java/com/bounteous/aem/compgenerator/utils/DialogUtils.java
@@ -1,0 +1,153 @@
+package com.bounteous.aem.compgenerator.utils;
+
+import com.bounteous.aem.compgenerator.Constants;
+import com.bounteous.aem.compgenerator.exceptions.GeneratorException;
+import com.bounteous.aem.compgenerator.models.GenerationConfig;
+import com.bounteous.aem.compgenerator.models.Property;
+import org.apache.commons.lang.StringUtils;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+
+import javax.xml.parsers.DocumentBuilderFactory;
+import java.util.List;
+import java.util.Objects;
+
+public class DialogUtils {
+
+    /**
+     * creates dialog xml by adding the properties in data-config json file.
+     *
+     * @param dialogType dialogType to dialog xml structure.
+     */
+    public static void createDialogXml(final GenerationConfig generationConfig, final String dialogType) {
+        String dialogPath = generationConfig.getCompDir() + "/" + dialogType;
+        try {
+            CommonUtils.createFolder(dialogPath);
+
+            Document doc = DocumentBuilderFactory.newInstance().newDocumentBuilder().newDocument();
+
+            Element rootElement = createDialogRoot(doc, generationConfig, dialogType);
+
+            List<Property> properties = generationConfig.getOptions().getProperties();
+            if (dialogType.equalsIgnoreCase(Constants.DIALOG_TYPE_GLOBAL)) {
+                properties = generationConfig.getOptions().getGlobalProperties();
+            } else if (dialogType.equalsIgnoreCase(Constants.DIALOG_TYPE_SHARED)) {
+                properties = generationConfig.getOptions().getSharedProperties();
+            }
+
+            if (properties != null && properties.size() > 0) {
+                Node currentNode = updateDefaultNodeStructure(doc, rootElement, dialogType);
+
+                properties.stream()
+                        .filter(Objects::nonNull)
+                        .map(property -> createPropertyNode(doc, property))
+                        .filter(Objects::nonNull)
+                        .forEach(a -> currentNode.appendChild(a));
+            }
+            doc.appendChild(rootElement);
+            XMLUtils.transformDomToFile(doc, dialogPath + "/" + Constants.FILENAME_CONTENT_XML);
+            System.out.println("Created : " + dialogPath + "/" + Constants.FILENAME_CONTENT_XML);
+        } catch (Exception e) {
+            throw new GeneratorException("Exception while creating Dialog xml : " + dialogPath);
+        }
+    }
+
+    private static Element createDialogRoot(Document document, GenerationConfig generationConfig, String dialogType) {
+        Element rootElement = XMLUtils.createRootElement(document);
+        rootElement.setAttribute(Constants.JCR_PRIMARY_TYPE, Constants.NT_UNSTRUCTURED);
+        rootElement.setAttribute(Constants.PROPERTY_SLING_RESOURCETYPE, Constants.RESOURCE_TYPE_DIALOG);
+        if (dialogType.equalsIgnoreCase(Constants.DIALOG_TYPE_GLOBAL)) {
+            rootElement.setAttribute(Constants.PROPERTY_JCR_TITLE,
+                    generationConfig.getTitle() + " (Global Properties)");
+        } else if (dialogType.equalsIgnoreCase(Constants.DIALOG_TYPE_SHARED)) {
+            rootElement.setAttribute(Constants.PROPERTY_JCR_TITLE,
+                    generationConfig.getTitle() + " (Shared Properties)");
+        } else {
+            rootElement.setAttribute(Constants.PROPERTY_JCR_TITLE, generationConfig.getTitle());
+        }
+        return rootElement;
+    }
+
+    /**
+     * adds a dialog property xml node with all input attr under the document.
+     *
+     * @param document
+     * @param property project object contains attributes.
+     * @return
+     */
+    private static Element createPropertyNode(Document document, Property property) {
+        try {
+            Element propertyNode = document.createElement(property.getField());
+
+            propertyNode.setAttribute(Constants.JCR_PRIMARY_TYPE, Constants.NT_UNSTRUCTURED);
+            propertyNode.setAttribute(Constants.PROPERTY_SLING_RESOURCETYPE, getSlingResourceType(property.getType()));
+            propertyNode.setAttribute(Constants.PROPERTY_CQ_MSM_LOCKABLE, "./" + property.getField());
+            propertyNode.setAttribute(Constants.PROPERTY_FIELDLABEL, property.getLabel());
+            propertyNode.setAttribute(Constants.PROPERTY_NAME, "./" + property.getField());
+
+            if (property.getAttributes() != null && property.getAttributes().size() > 0) {
+                property.getAttributes()
+                        .entrySet()
+                        .stream()
+                        .forEach(entry -> propertyNode.setAttribute(entry.getKey(), entry.getValue()));
+            }
+            return propertyNode;
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    /**
+     * builds default node structure of dialog xml in the document passed in based on dialogType.
+     *
+     * @param document
+     * @param root
+     * @param dialogType
+     * @return
+     */
+    private static Node updateDefaultNodeStructure(Document document, Element root, String dialogType) {
+        Element containerElement = document.createElement("content");
+        containerElement.setAttribute(Constants.JCR_PRIMARY_TYPE, Constants.NT_UNSTRUCTURED);
+        containerElement.setAttribute(Constants.PROPERTY_SLING_RESOURCETYPE, Constants.RESOURCE_TYPE_CONTAINER);
+
+        Element layoutElement1 = document.createElement("layout");
+        layoutElement1.setAttribute(Constants.JCR_PRIMARY_TYPE, Constants.NT_UNSTRUCTURED);
+        layoutElement1.setAttribute(Constants.PROPERTY_SLING_RESOURCETYPE, Constants.RESOURCE_TYPE_FIXEDCOLUMNS);
+        layoutElement1.setAttribute("margin", "{Boolean}false");
+
+        Element columnElement = document.createElement("column");
+        columnElement.setAttribute(Constants.JCR_PRIMARY_TYPE, Constants.NT_UNSTRUCTURED);
+        columnElement.setAttribute(Constants.PROPERTY_SLING_RESOURCETYPE, Constants.RESOURCE_TYPE_CONTAINER);
+
+        Node containerNode = root.appendChild(containerElement);
+
+        containerNode.appendChild(layoutElement1);
+        return containerNode
+                .appendChild(createUnStructuredNode(document, "items"))
+                .appendChild(columnElement)
+                .appendChild(createUnStructuredNode(document, "items"));
+    }
+
+    private static Node createUnStructuredNode(Document document, String nodeName) {
+        Element element = document.createElement(nodeName);
+        element.setAttribute(Constants.JCR_PRIMARY_TYPE, Constants.NT_UNSTRUCTURED);
+        return element;
+    }
+
+    private static String getSlingResourceType(String type){
+        if(StringUtils.isNotBlank(type)){
+            if (StringUtils.equalsIgnoreCase("text", type)) {
+                return Constants.RESOURCE_TYPE_TEXTFIELD;
+            } else if (StringUtils.equalsIgnoreCase("number", type)) {
+                return Constants.RESOURCE_TYPE_NUMBER;
+            } else if (StringUtils.equalsIgnoreCase("checkbox", type)) {
+                return Constants.RESOURCE_TYPE_CHECKBOX;
+            }
+        }
+        return null;
+    }
+
+
+}
